add user base order get, getAll, create, detete

+++ api/controllers/orders.js
@@ -2,18 +2,54 @@ const sequelizeOrder = require('../sequelize-models/Order')
 const sequelizeUserOrderMapping = require('../sequelize-models/UserOrderMapping')
 
 exports.orders_get_all = async (req, res, next) => {
-    console.log("orders", req.body);
+    console.log("orders_get", req.body);
+
+    // console.log("one", req.user.id);
+
+    const current_user_id = req.user.id
+    console.log(current_user_id)
+
+
+    const userOrders = [];
+    const findUserId = await sequelizeUserOrderMapping.findAll({
+        where: {
+            user_id: current_user_id,
+        },
+        attributes: ["order_id"]
+    })
+    // console.log("find", findUserId.user_order_mapping)
+    findUserId.map((val) => {
+        // console.log('Value', val.order_id);
+        userOrders.push(val.order_id)
+    })
+    console.log("list", userOrders);
+
+    // console.log("id", req.params.id)
+    // console.log("type", typeof (req.params.id))
+    // console.log("type", typeof parseInt(req.params.id))
+
 
     try {
-        const orderAll = await sequelizeOrder.findAll({
+        if (userOrders) {
+            console.log("two")
+            const order = await sequelizeOrder.findAll({
+                attributes: ['product_id', 'quantity'],
+                where: {
+                    id: userOrders
+                }
+            })
+            // if (!order) {
+            //     return res.status(404).send({ message: "Order is not found for deleting" });
+            // }
+            res.json({
+                message: "Find successfully", order
+            })
 
-            attributes: ['id', 'product', 'quantity']
+        } else {
+            console.log("three")
+            return res.status(404).send({ message: "Not found!!!" });
+        }
 
-        })
-
-        res.json({
-            message: orderAll
-        })
     }
     catch (err) {
         console.log(err)
@@ -21,23 +57,80 @@ exports.orders_get_all = async (req, res, next) => {
             error: err
         })
     }
+
+
+
+    // console.log("orders", req.body);
+
+    // try {
+    //     const orderAll = await sequelizeOrder.findAll({
+
+    //         attributes: ['id', 'product_id', 'quantity']
+
+    //     })
+
+    //     res.json({
+    //         message: orderAll
+    //     })
+    // }
+    // catch (err) {
+    //     console.log(err)
+    //     res.status(500).json({
+    //         error: err
+    //     })
+    // }
 };
 
 exports.orders_get_order = async (req, res, next) => {
-    console.log("order_id", req.params);
+    console.log("orders_get", req.body);
+
+    // console.log("one", req.user.id);
+
+    const current_user_id = req.user.id
+    console.log(current_user_id)
+
+
+    const userOrders = [];
+    const findUserId = await sequelizeUserOrderMapping.findAll({
+        where: {
+            user_id: current_user_id,
+        },
+        attributes: ["order_id"]
+    })
+    // console.log("find", findUserId.user_order_mapping)
+    findUserId.map((val) => {
+        // console.log('Value', val.order_id);
+        userOrders.push(val.order_id)
+    })
+    console.log("list", userOrders);
+
+    // console.log("id", req.params.id)
+    // console.log("type", typeof (req.params.id))
+    // console.log("type", typeof parseInt(req.params.id))
+
 
     try {
-        const { id } = req.params;
-        const orderId = await sequelizeOrder.findAll({
-            attributes: ['id', 'product', 'quantity'],
-            where: {
-                id
-            }
-        })
+        if (userOrders.includes(parseInt(req.params.id))) {
+            console.log("two")
+            const { id } = req.params;
+            const order = await sequelizeOrder.findOne({
+                attributes: ['product_id', 'quantity'],
+                where: {
+                    id
+                }
+            })
+            // if (!order) {
+            //     return res.status(404).send({ message: "Order is not found for deleting" });
+            // }
+            res.json({
+                message: "Find successfully", order
+            })
 
-        res.json({
-            message: "orderId find successfully", orderId
-        })
+        } else {
+            console.log("three")
+            return res.status(404).send({ message: "Not found!!!" });
+        }
+
     }
     catch (err) {
         console.log(err)
@@ -45,12 +138,35 @@ exports.orders_get_order = async (req, res, next) => {
             error: err
         })
     }
+
+
+    // console.log("order_id", req.params);
+
+    // try {
+    //     const { id } = req.params;
+    //     const orderId = await sequelizeOrder.findOne({
+    //         attributes: ['id', 'product_id', 'quantity'],
+    //         where: {
+    //             id
+    //         }
+    //     })
+
+    //     res.json({
+    //         message: "orderId find successfully", orderId
+    //     })
+    // }
+    // catch (err) {
+    //     console.log(err)
+    //     res.status(500).json({
+    //         error: err
+    //     })
+    // }
 };
 
 exports.orders_create_order = async (req, res, next) => {
 
     console.log("orders_create", req.body);
-    console.log("two", req.user);
+    // console.log("two", req.user);
     console.log("three", req.user.id);
 
     const user_id = req.user.id
@@ -62,7 +178,7 @@ exports.orders_create_order = async (req, res, next) => {
             product_id,
             quantity
         })
-        console.log(newOrder)
+        // console.log(newOrder)
         const newOrderId = newOrder.id
         console.log("newOrderID", newOrder.id)
 
@@ -89,19 +205,60 @@ exports.orders_create_order = async (req, res, next) => {
 exports.orders_delete_order = async (req, res, next) => {
     console.log("orders_deleted", req.body);
 
+    // console.log("one", req.user.id);
+
+    const current_user_id = req.user.id
+    // console.log(current_user_id)
+
+
+    const userOrders = [];
+    const findUserId = await sequelizeUserOrderMapping.findAll({
+        where: {
+            user_id: current_user_id,
+        },
+        attributes: ["order_id"]
+    })
+    // console.log("find", findUserId.user_order_mapping)
+    findUserId.map((val) => {
+        console.log('Value', val.order_id);
+        userOrders.push(val.order_id)
+    })
+    console.log("list", userOrders);
+
+
     try {
-        const { id } = req.body;
-        const order = await sequelizeOrder.destroy({
-            where: {
-                id
-            }
-        })
-        if (!order) {
+        if (userOrders.includes(req.body.id)) {
+            // console.log("two")
+            const { id } = req.body;
+            const order = await sequelizeOrder.destroy({
+                where: {
+                    id
+                }
+            })
+            // if (!order) {
+            //     return res.status(404).send({ message: "Order is not found for deleting" });
+            // }
+            res.json({
+                message: "Order deleted successfully"
+            })
+
+        } else {
+            // console.log("three")
             return res.status(404).send({ message: "Order is not found for deleting" });
         }
-        res.json({
-            message: "Order deleted successfully"
-        })
+
+        // const { id } = req.body;
+        // const order = await sequelizeOrder.destroy({
+        //     where: {
+        //         id
+        //     }
+        // })
+        // if (!order) {
+        //     return res.status(404).send({ message: "Order is not found for deleting" });
+        // }
+        // res.json({
+        //     message: "Order deleted successfully"
+        // })
     }
     catch (err) {
         console.log(err)
